requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.4.0
ipykernel==6.5.1
ipython==7.30.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.2
tornado==6.1
traitlets==5.1.1
virtualenv==20.10.0
wcwidth==0.2.5
